added some more to the OrbitalStrikeSmall
